separe functions in diferentt files
